Localize plugin string format usages
